src/db_utils.py
# ─────────────────────────────────────────────
# db_utils.py — PostgreSQL connection & chat history persistence
# ─────────────────────────────────────────────
import os
from dotenv import load_dotenv
from langchain_postgres import PostgresChatMessageHistory
import psycopg

load_dotenv()

# Load connection string from environment
DATABASE_URL = os.getenv("DATABASE_URL")

def get_postgres_history(session_id: str, table_name: str = "chat_history"):
    """
    Returns a PostgresChatMessageHistory object for the given session.
    If DATABASE_URL is missing, it returns None.
    """
    if not DATABASE_URL:
        print("DEBUG: DATABASE_URL not found in environment. Persistent memory disabled.")
        return None

    try:
        # Initialize a psycopg connection correctly for newer langchain-postgres
        conn = psycopg.connect(DATABASE_URL)
        
        # Initialize the history object using positional-only arguments for table_name and session_id
        # and pasing the sync_connection object.
        history = PostgresChatMessageHistory(
            table_name,
            session_id,
            sync_connection=conn
        )
        
        # Ensure the table exists
        PostgresChatMessageHistory.create_tables(conn, table_name)
        
        print(f"DEBUG: Successfully connected to PostgreSQL for session: {session_id}")
        return history
    except Exception as e:
        print(f"ERROR: Failed to connect to PostgreSQL: {e}")
        return None
